requirements.txt
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
Brotli==1.1.0
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
executing==2.0.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
ipython==8.16.1
jedi==0.19.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
model-bakery==1.16.0
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
pytest-testdox==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
wcwidth==0.2.8
whitenoise==6.6.0
